fixed the gui, hopefully once and for all

+++ src/main/java/subsym/genetics/adultselection/OverProduction.java
@@ -10,7 +10,7 @@ public class OverProduction implements AdultSelection {
 
   public double overProductionRate ;
 
-  public OverProduction(int overProductionRate) {
+  public OverProduction(double overProductionRate) {
     this.overProductionRate = overProductionRate;
   }
